python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
hdbcli>=2.19.0
anthropic>=0.40.0
openai>=1.55.0
google-generativeai>=0.8.0
